+++ langchain-interpreter/src/afm/resources/chat-ui.html
